refactor: relocate util funcs to new files to be compatible with #4834
